pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "claude-skill-check"
version = "0.1.0"
description = "Linter for Claude Code SKILL.md files. Validates YAML frontmatter, required fields, description length, and common secret patterns."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Mukunda Katta", email = "mukunda.vjcs6@gmail.com" }]
keywords = ["claude", "claude-code", "skill", "anthropic", "linter", "validator", "cli"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.urls]
Homepage = "https://github.com/MukundaKatta/claude-skill-check"
Issues = "https://github.com/MukundaKatta/claude-skill-check/issues"
Source = "https://github.com/MukundaKatta/claude-skill-check"

[project.scripts]
claude-skill-check = "claude_skill_check.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_skill_check"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests", "/README.md", "/LICENSE"]
